pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "metachrome"
dynamic = ["version"]
description = "Metaphase Chromosome Analysis with Napari"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Md Abdul Kader Sagar", email = "sagarm2@nih.gov"}
]
maintainers = [
    {name = "Md Abdul Kader Sagar", email = "sagarm2@nih.gov"}
]
keywords = ["chromosome", "analysis", "napari", "bioinformatics", "image-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.8"
dependencies = [
    "napari>=0.4.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-image>=0.18.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "magicgui>=0.2.0",
    "qtpy>=1.9.0",
    "superqt>=0.2.0",
    "cellpose>=2.0.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
]

[project.scripts]
metachrome = "metachrome.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/metachrome"
Repository = "https://github.com/yourusername/metachrome"
Documentation = "https://github.com/yourusername/metachrome#readme"
"Bug Tracker" = "https://github.com/yourusername/metachrome/issues"

[tool.setuptools]
packages = ["metachrome"]

[tool.setuptools.package-data]
metachrome = ["*.py"]

[tool.setuptools_scm]
write_to = "metachrome/_version.py"
